moved reusable function
