+++ public/src/ts/menu/saveAs.ts
@@ -1,0 +1,7 @@
+import { graph } from '../graph.js'
+import { save } from '../util.js';
+
+export function savePNG(): void {
+  const filename = "myDiagram.png";
+  
+}
