+++ public/src/ts/util.ts
@@ -2,22 +2,40 @@ import { graph } from "./graph.js"
 import { selected_links } from "./tools/LinkButton.js"
 import { selected_premises } from "./tools/CombinePremise.js"
 
+export function save(data: string, filetype: string, filename: string): void {
+	const file = new Blob([data], {type: filetype});    
+  if (window.navigator.msSaveOrOpenBlob) {// IE10+
+    window.navigator.msSaveOrOpenBlob(file, filename);
+  } else { // Others
+		const a = document.createElement("a");
+		const url = URL.createObjectURL(file);
+		a.href = url;
+		a.download = filename;
+		document.body.appendChild(a);
+		a.click();
+		setTimeout(function() {
+			document.body.removeChild(a);
+			window.URL.revokeObjectURL(url);  
+		}, 0); 
+	}
+}
+
 export function calcHeight(num_lines:number) {
-    return 16 + 13 * num_lines;
+  return 16 + 13 * num_lines;
 }
 
 graph.on('remove', function(cell) {
-    if (!cell.isLink()) {
-        //something other than a link was removed
-        //clear queues
-       // have to be careful here, since this event can be triggered by removal of two
-       // component shapes when making a dependent premise, so we only want to do this
-       // when queue is half full
-        if (selected_links.length === 1) {
-            selected_links.pop();
-        }
-        if (selected_premises.length == 1 ) {
-            selected_premises.pop();
-        }
+  if (!cell.isLink()) {
+    //something other than a link was removed
+    //clear queues
+    // have to be careful here, since this event can be triggered by removal of two
+    // component shapes when making a dependent premise, so we only want to do this
+    // when queue is half full
+    if (selected_links.length === 1) {
+      selected_links.pop();
     }
+    if (selected_premises.length == 1 ) {
+      selected_premises.pop();
+    }
+  }
 })
